requirements.txt
cachelib==0.9.0
click==7.1.2
Flask==1.1.2
Flask-Caching==1.11.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.19.5
psycopg2-binary==2.9.3
scipy==1.5.4
Werkzeug==1.0.1
WTForms==3.0.1
Flask-WTF==1.0.1